implement gzserver and gzstorage
