+++ gzserver_test.go
@@ -11,11 +11,174 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"net/url"
+	"strings"
 	"testing"
 
 	"github.com/rs/zerolog"
+
 	"github.com/stretchr/testify/assert"
 )
+
+func TestGzServerTranslateRequestToExpectation_SimpleRequest_AllFieldsTranslated(t *testing.T) {
+	request, err := http.NewRequest("POST", "https://www.host.com/a/b?foo=bar#fr", strings.NewReader("body text"))
+	if err != nil {
+		t.Fatal(err)
+	}
+	request.Header.Add("h1", "hv1")
+
+	exp := translateRequestToExpectation(request)
+
+	assert.NotNil(t, exp)
+	assert.Equal(t, "POST", exp.Method)
+	assert.Equal(t, "/a/b?foo=bar#fr", exp.Path)
+	assert.Equal(t, "body text", exp.Body)
+	assert.NotNil(t, exp.Headers)
+	assert.Equal(t, 1, len(*exp.Headers))
+	assert.Equal(t, "hv1", (*exp.Headers)["H1"])
+}
+
+func TestGzServerTranslateHTTPHeadersToExpHeaders_TwoHeaders_HeadersTranslated(t *testing.T) {
+	header := http.Header{}
+	header.Add("h1", "hv1")
+	header.Add("h1", "hv2")
+
+	expHeaders := translateHTTPHeadersToExpHeaders(header)
+	assert.NotNil(t, expHeaders)
+	assert.Equal(t, 1, len(*expHeaders))
+	assert.Equal(t, "hv1,hv2", (*expHeaders)["H1"])
+}
+
+func TestGzServerControllerStringPassesFilter_EmptyFilter_True(t *testing.T) {
+	assert.True(t, controllerStringPassesFilter("abc", ""))
+}
+
+func TestGzServerControllerStringPassesFilter_ExistingSubstring_True(t *testing.T) {
+	assert.True(t, controllerStringPassesFilter("abc", "ab"))
+}
+
+func TestGzServerControllerStringPassesFilter_ExistingRegex_True(t *testing.T) {
+	assert.True(t, controllerStringPassesFilter("abc", ".b."))
+}
+
+func TestGzServerControllerStringPassesFilter_NotExistingSubstring_False(t *testing.T) {
+	assert.False(t, controllerStringPassesFilter("abc", "zz"))
+}
+
+func TestGzServerControllerStringPassesFilter_NotExistingRegex_False(t *testing.T) {
+	assert.False(t, controllerStringPassesFilter("abc", ".z."))
+}
+
+func TestGzServerControllerStringPassesFilter_MultilineBody_True(t *testing.T) {
+	assert.True(t, controllerStringPassesFilter("a\nb", "a.b"))
+}
+
+func TestGzServerControllerRequestPassFilter_EmptyRequestEmptyFilter_True(t *testing.T) {
+	assert.True(t, controllerRequestPassesFilter(
+		&ExpectationRequest{},
+		&ExpectationRequest{}))
+}
+
+func TestGzServerControllerRequestPassFilter_MethodsAreEq_True(t *testing.T) {
+	assert.True(t, controllerRequestPassesFilter(
+		&ExpectationRequest{Method: "POST"},
+		&ExpectationRequest{Method: "POST"}))
+}
+
+func TestGzServerControllerRequestPassFilter_PathsAreEq_True(t *testing.T) {
+	assert.True(t, controllerRequestPassesFilter(
+		&ExpectationRequest{Path: "/path"},
+		&ExpectationRequest{Path: "/path"}))
+}
+
+func TestGzServerControllerRequestPassFilter_MethodsNotEqAndPathsAreEq_False(t *testing.T) {
+	assert.False(t, controllerRequestPassesFilter(
+		&ExpectationRequest{Method: "GET", Path: "/path"},
+		&ExpectationRequest{Method: "POST", Path: "/path"}))
+}
+
+func TestGzServerControllerRequestPassFilter_HeadersAreEq_True(t *testing.T) {
+	assert.True(t, controllerRequestPassesFilter(
+		&ExpectationRequest{Headers: &Headers{"h1": "hv1"}},
+		&ExpectationRequest{Headers: &Headers{"h1": "hv1"}}))
+}
+
+func TestGzServerControllerRequestPassFilter_HeaderNotEq_False(t *testing.T) {
+	result := controllerRequestPassesFilter(
+		&ExpectationRequest{Headers: &Headers{"h1": "hv1"}},
+		&ExpectationRequest{Headers: &Headers{"h2": "hv2"}})
+	assert.False(t, result)
+}
+
+func TestGzServerControllerRequestPassFilter_HeaderValueNotEq_False(t *testing.T) {
+	assert.False(t, controllerRequestPassesFilter(
+		&ExpectationRequest{Headers: &Headers{"h1": "hv1"}},
+		&ExpectationRequest{Headers: &Headers{"h1": "hv2"}}))
+}
+
+func TestGzServerControllerRequestPassFilter_NoHeaderinReq_False(t *testing.T) {
+	assert.False(t, controllerRequestPassesFilter(
+		&ExpectationRequest{},
+		&ExpectationRequest{Headers: &Headers{"h2": "hv2"}}))
+}
+
+func TestGzServerControllerRequestPassFilter_NoHeaderInFilter_True(t *testing.T) {
+	assert.True(t, controllerRequestPassesFilter(
+		&ExpectationRequest{Headers: &Headers{"h1": "hv1"}},
+		&ExpectationRequest{}))
+}
+
+func TestGzServerControllerRequestPassFilter_BodysEq_True(t *testing.T) {
+	assert.True(t, controllerRequestPassesFilter(
+		&ExpectationRequest{Body: "body"},
+		&ExpectationRequest{Body: "body"}))
+}
+
+func TestGzServerControllerCreateHTTPRequestWithHeaders(t *testing.T) {
+	expReq := &ExpectationRequest{Method: "GET", Path: "/request", Headers: &Headers{"h_req": "hv_req"}}
+	expFwd := &ExpectationForward{Scheme: "https", Host: "localhost_fwd", Headers: &Headers{"h_req": "hv_fwd", "h_fwd": "hv_fwd"}}
+	httpReq := controllerCreateHTTPRequest(expReq, expFwd)
+	assert.NotNil(t, httpReq)
+	assert.Equal(t, expReq.Method, httpReq.Method)
+	assert.Equal(t, expFwd.Host, httpReq.Host)
+	assert.Equal(t, fmt.Sprintf("%s://%s%s", expFwd.Scheme, expFwd.Host, expReq.Path), httpReq.URL.String())
+	assert.Equal(t, "hv_fwd", httpReq.Header.Get("h_req"))
+	assert.Equal(t, "hv_fwd", httpReq.Header.Get("h_fwd"))
+}
+
+func TestGzServerControllerCreateHTTPRequestHostRewrite(t *testing.T) {
+	expReq := &ExpectationRequest{Method: "GET", Path: "/request"}
+	expFwd := &ExpectationForward{Scheme: "https", Host: "localhost_fwd", Headers: &Headers{"Host": "fwd_host"}}
+	httpReq := controllerCreateHTTPRequest(expReq, expFwd)
+	assert.NotNil(t, httpReq)
+	assert.Equal(t, "fwd_host", httpReq.Host)
+	assert.Equal(t, fmt.Sprintf("%s://%s%s", expFwd.Scheme, expFwd.Host, expReq.Path), httpReq.URL.String())
+}
+
+// Integration tests
+
+func httpNewRequestMust(method, url string, body io.Reader) *http.Request {
+	req, err := http.NewRequest(method, url, body)
+	if err != nil {
+		panic(err)
+	}
+	return req
+}
+
+func TestGzServer_NoExpectations_ReturnEmptyList(t *testing.T) {
+	//Arrange
+	server := gzServer{logLevel: zerolog.DebugLevel}
+	server.storage.init()
+
+	r := httpNewRequestMust("GET", "/gozzmock/get_expectations", nil)
+	w := httptest.NewRecorder()
+
+	//Act
+	server.get(w, r)
+
+	//Assert
+	assert.Equal(t, http.StatusNotImplemented, w.Code)
+	assert.Equal(t, "{}", w.Body.String())
+}
 
 func jsonMarshalMust(v interface{}) []byte {
 	encoded, err := json.Marshal(v)
@@ -26,26 +189,17 @@ func jsonMarshalMust(v interface{}) []byte {
 	return encoded
 }
 
-func httpNewRequestMust(method, url string, body io.Reader) *http.Request {
-	req, err := http.NewRequest(method, url, body)
-	if err != nil {
-		panic(err)
-	}
-	return req
-}
-
-func (context *Context) getExpectations(t *testing.T) *bytes.Buffer {
+/*
+func (context *gzServerMock) getExpectations(t *testing.T) *bytes.Buffer {
 	handlerGetExpectations := http.HandlerFunc(context.HandlerGetExpectations)
-	req := httpNewRequestMust("GET", "/gozzmock/get_expectations", nil)
 
-	httpTestResponseRecorder := httptest.NewRecorder()
 	handlerGetExpectations.ServeHTTP(httpTestResponseRecorder, req)
 	assert.Equal(t, http.StatusOK, httpTestResponseRecorder.Code)
 
 	return httpTestResponseRecorder.Body
 }
 
-func (context *Context) addExpectation(t *testing.T, exp Expectation) *bytes.Buffer {
+func (context *gzServerMock) addExpectation(t *testing.T, exp Expectation) *bytes.Buffer {
 	handlerAddExpectation := http.HandlerFunc(context.HandlerAddExpectation)
 
 	expJSON := jsonMarshalMust(exp)
@@ -60,7 +214,7 @@ func (context *Context) addExpectation(t *testing.T, exp Expectation) *bytes.Buf
 	return context.getExpectations(t)
 }
 
-func (context *Context) removeExpectation(t *testing.T, expKey string) *bytes.Buffer {
+func (context *gzServerMock) removeExpectation(t *testing.T, expKey string) *bytes.Buffer {
 	handlerRemoveExpectation := http.HandlerFunc(context.HandlerRemoveExpectation)
 
 	expRemoveJSON := jsonMarshalMust(ExpectationRemove{Key: expKey})
@@ -74,13 +228,13 @@ func (context *Context) removeExpectation(t *testing.T, expKey string) *bytes.Bu
 
 	return context.getExpectations(t)
 }
+*/
 
 func TestHandlerNoExpectations(t *testing.T) {
 	// Arrange
-	context := Context{
-		logLevel: zerolog.DebugLevel,
-		storage:  ControllerCreateStorage()}
-	handlerDefault := http.HandlerFunc(context.HandlerDefault)
+	server := gzServer{logLevel: zerolog.DebugLevel}
+	server.storage.init()
+	handlerDefault := http.HandlerFunc(server.root)
 	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Write([]byte("response from test server"))
 	}))
@@ -101,9 +255,9 @@ func TestHandlerNoExpectations(t *testing.T) {
 
 func TestHandlerAddAndRemoveExpectation(t *testing.T) {
 	// Arrange
-	context := Context{
-		logLevel: zerolog.DebugLevel,
-		storage:  ControllerCreateStorage()}
+	context := gzServer{
+		logLevel: zerolog.DebugLevel}
+	context.storage.init()
 	expectedExp := Expectation{Key: "k"}
 	expectedExps := Expectations{expectedExp.Key: expectedExp}
 	expsjson := jsonMarshalMust(expectedExps)
@@ -119,9 +273,8 @@ func TestHandlerAddAndRemoveExpectation(t *testing.T) {
 
 func TestHandlerAddTwoExpectations(t *testing.T) {
 	// Arrange
-	context := Context{
-		logLevel: zerolog.DebugLevel,
-		storage:  ControllerCreateStorage()}
+	context := gzServer{logLevel: zerolog.DebugLevel}
+	context.storage.init()
 	handlerDefault := http.HandlerFunc(context.HandlerDefault)
 	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		dumpRequest(r)
@@ -165,9 +318,8 @@ func TestHandlerAddTwoExpectations(t *testing.T) {
 
 func TestHandlerGetExpectations(t *testing.T) {
 	// Arrange
-	context := Context{
-		logLevel: zerolog.DebugLevel,
-		storage:  ControllerCreateStorage()}
+	context := gzServer{logLevel: zerolog.DebugLevel}
+	context.storage.init()
 	handlerGetExpectations := http.HandlerFunc(context.HandlerGetExpectations)
 	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Write([]byte("response from test server"))
@@ -198,9 +350,8 @@ func TestHandlerGetExpectations(t *testing.T) {
 
 func TestHandlerStatus(t *testing.T) {
 	// Arrange
-	context := Context{
-		logLevel: zerolog.DebugLevel,
-		storage:  ControllerCreateStorage()}
+	context := gzServer{logLevel: zerolog.DebugLevel}
+	context.storage.init()
 	handlerStatus := http.HandlerFunc(context.HandlerStatus)
 	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Write([]byte("response from test server"))
@@ -221,9 +372,8 @@ func TestHandlerStatus(t *testing.T) {
 
 func TestHandlerForwardValidatrHeaders(t *testing.T) {
 	// Arrange
-	context := Context{
-		logLevel: zerolog.InfoLevel,
-		storage:  ControllerCreateStorage()}
+	context := gzServer{logLevel: zerolog.DebugLevel}
+	context.storage.init()
 	handlerDefault := http.HandlerFunc(context.HandlerDefault)
 	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Write([]byte(fmt.Sprint(r.Host)))
@@ -263,9 +413,8 @@ func writeCompressedMessage(w http.ResponseWriter, message []byte) {
 
 func TestHandlerForwardReturnsGzip(t *testing.T) {
 	// Arrange
-	context := Context{
-		logLevel: zerolog.DebugLevel,
-		storage:  ControllerCreateStorage()}
+	context := gzServer{logLevel: zerolog.DebugLevel}
+	context.storage.init()
 	handlerDefault := http.HandlerFunc(context.HandlerDefault)
 	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		writeCompressedMessage(w, []byte("response from test server"))
@@ -307,9 +456,8 @@ func TestHandlerForwardReturnsGzip(t *testing.T) {
 
 func TestHandlerRespondsWithJsTemplate(t *testing.T) {
 	// Arrange
-	context := Context{
-		logLevel: zerolog.DebugLevel,
-		storage:  ControllerCreateStorage()}
+	context := gzServer{logLevel: zerolog.DebugLevel}
+	context.storage.init()
 
 	handlerDefault := http.HandlerFunc(context.HandlerDefault)
 	jsTemplate := `"123".length`
